+++ dynamite/config_reader.py
@@ -123,6 +123,9 @@ class Configuration(object):
     reset_logging : bool
         if False: use the calling application's logging settings
         if True: set logging to Dynamite defaults
+    reset_existing_output : bool
+        if False: do not touch existing data in the output directory tree
+        if True: rebuild the output directory tree and delete existing data
 
     Raises
     ------
@@ -144,7 +147,8 @@ class Configuration(object):
     thresh_chi2_abs = 'threshold_del_chi2_abs'
     thresh_chi2_scaled = 'threshold_del_chi2_as_frac_of_sqrt2nobs'
 
-    def __init__(self, filename=None, silent=None, reset_logging=False):
+    def __init__(self, filename=None, silent=None, reset_logging=False,
+                 reset_existing_output=False):
         if reset_logging is True:
             DynamiteLogging()
             self.logger = logging.getLogger(f'{__name__}.{__class__.__name__}')
@@ -196,7 +200,13 @@ class Configuration(object):
             raise
         self.settings.add('io_settings', self.params['io_settings'])
         logger.debug('io_settings assigned to Settings object')
+        out_dir = self.settings.io_settings['output_directory']
+        if reset_existing_output:
+            if os.path.isdir(out_dir):
+                shutil.rmtree(out_dir)
+                self.logger.info(f'Output directory tree {out_dir} removed.')
         self.make_output_directory_tree()
+        self.logger.info(f'Output directory tree: {out_dir}.')
 
         for key, value in self.params.items(): # walk through file contents...
 
@@ -295,7 +305,7 @@ class Configuration(object):
 
                     # add component to system
                     c.validate()
-                    parset = {c.get_parname(p.name):p.value \
+                    parset = {c.get_parname(p.name):p.raw_value \
                               for p in c.parameters}
                     if not c.validate_parset(parset):
                         text = f'{c.name}: invalid parameters {parset}'
@@ -374,6 +384,10 @@ class Configuration(object):
 
             elif key == 'weight_solver_settings':
                 logger.info('weight_solver_settings...')
+                if 'reattempt_failures' not in value:
+                    value['reattempt_failures'] = True
+                if value['reattempt_failures']:
+                    logger.info('Will attempt to recover partially run models.')
                 logger.debug(f'weight_solver_settings: {tuple(value.keys())}')
                 self.settings.add('weight_solver_settings', value)
 
@@ -389,20 +403,17 @@ class Configuration(object):
                     pass
                 if value['ncpus']=='all_available':
                     value['ncpus'] = self.get_n_cpus()
-                if not silent:
-                    logger.info(f"... using {value['ncpus']} CPUs "
-                                "for orbit integration.")
+                logger.info(f"... using {value['ncpus']} CPUs "
+                             "for orbit integration.")
                 if 'ncpus_weights' not in value:
                     value['ncpus_weights'] = value['ncpus']
                 elif value['ncpus_weights'] == 'all_available':
                     value['ncpus_weights'] = self.get_n_cpus()
-                if not silent:
-                    logger.info(f"... using {value['ncpus_weights']} CPUs "
-                                "for weight solving.")
+                logger.info(f"... using {value['ncpus_weights']} CPUs "
+                            "for weight solving.")
                 if 'modeliterator' not in value:
                     value['modeliterator'] = 'ModelInnerIterator'
-                if not silent:
-                    logger.info(f"... using iterator {value['modeliterator']}.")
+                logger.debug(f"... using iterator {value['modeliterator']}.")
                 if 'orblibs_in_parallel' not in value:
                     value['orblibs_in_parallel'] = False
                 logger.debug("... integrate orblibs in parallel: "
@@ -416,7 +427,7 @@ class Configuration(object):
                 raise ValueError(text)
 
         self.system.validate() # now also adds the right parameter sformat
-        parset = {p.name:p.value for p in self.system.parameters}
+        parset = {p.name:p.raw_value for p in self.system.parameters}
         if not self.system.validate_parset(parset):
             text = f'Invalid system parameters {parset}'
             self.logger.error(text)
@@ -440,7 +451,7 @@ class Configuration(object):
         logger.info('Instantiated AllModels object')
         logger.debug(f'AllModels:\n{self.all_models.table}')
 
-        self.backup_config_file(reset=False)
+        # self.backup_config_file(reset=False)
 
     def get_n_cpus(self):
         """"
@@ -490,7 +501,7 @@ class Configuration(object):
             - ``BayesLOSVD``, then n_obs = n_LOSVD_bins * number_spatial_bins
 
         This returns the sum of (2 * n_obs) for all kinematic sets. Take
-        kinemtics from the ``TriaxialVisibleComponent`` of the system.
+        kinematics from the ``TriaxialVisibleComponent`` of the system.
 
         Returns
         -------
@@ -620,7 +631,7 @@ class Configuration(object):
             for f in glob.glob(f'{output_dir}*'):
                 if os.path.isfile(f):
                     os.remove(f)
-            self.backup_config_file()
+            # self.backup_config_file()
             self.logger.info(f'Removed files in {output_dir}.')
         else:
             all_models_file = self.settings.io_settings['output_directory'] \
@@ -677,7 +688,7 @@ class Configuration(object):
         self.remove_existing_all_models_file()
         if create_tree:
             self.make_output_directory_tree()
-            self.backup_config_file()
+            # self.backup_config_file()
 
     def make_output_directory_tree(self):
         """
@@ -707,6 +718,34 @@ class Configuration(object):
         else:
             self.logger.debug(f'Using existing plots directory {plot_dir}.')
 
+    def copy_config_file(self, dest_directory, clean=True):
+        """
+        Copy config file to dest_directory.
+
+        Creates a copy of the config file, intended to add it to the directory
+        holding the model results. The file date will be preserved if possible.
+
+        Parameters
+        ----------
+        dest_directory : str, mandatory
+            The directory the config file will be copied to.
+        clean : bool, optional
+            If True, all *.yaml files in dest_directory will be deleted before
+            copying. Default is True.
+        """
+        if dest_directory[-1] != '/':
+            dest_directory += '/'
+        if clean:
+            del_files = glob.iglob(f'{dest_directory}*.yaml')
+            for f_name in del_files:
+                if os.path.isfile(f_name):
+                    os.remove(f_name)
+            self.logger.debug(f'{dest_directory}*.yaml files deleted.')
+        shutil.copy2(self.config_file_name, dest_directory)
+        self.logger.info('Config file copied to '
+                         f'{dest_directory}{self.config_file_name}.')
+
+
     def backup_config_file(self, reset=False, keep=None, delete_other=False):
         """
         Copy the config file to the output directory.
@@ -714,6 +753,8 @@ class Configuration(object):
         A running index of the format _xxx will be appended to the base file
         name to keep track of earlier config files (config_000.yaml,
         config_001.yaml, config_002.yaml, etc...)
+
+        This method is not used in standard DYNAMITE and provided as a utility.
 
         Parameters
         ----------
@@ -847,15 +888,15 @@ class Configuration(object):
             if issubclass(type(c), physys.DarkComponent) \
                 and not isinstance(c, physys.Plummer):
             # Check allowed dm halos in legacy mode
-                if type(c) not in [physys.NFW, physys.Hernquist,
+                if type(c) not in [physys.NFW, physys.NFW_m200_c,
+                                   physys.Hernquist,
                                    physys.TriaxialCoredLogPotential,
                                    physys.GeneralisedNFW]:
-                    self.logger.error(f'DM Halo needs to be of type NFW, '
-                                      f'Hernquist, TriaxialCoredLogPotential, '
-                                      f'or GeneralisedNFW, not {type(c)}')
-                    raise ValueError(f'DM Halo needs to be of type NFW, '
-                                     f'Hernquist, TriaxialCoredLogPotential, '
-                                     f'or GeneralisedNFW, not {type(c)}')
+                    text = 'DM Halo needs to be of type NFW, NFW_m200_c, ' \
+                           'Hernquist, TriaxialCoredLogPotential, ' \
+                           f'or GeneralisedNFW, not {type(c)}'
+                    self.logger.error(text)
+                    raise ValueError(text)
 
         gen_type = self.settings.parameter_space_settings["generator_type"]
         allowed_types = ['GridWalk', 'LegacyGridSearch', 'FullGrid']