+++ docs/more_info/changelog.rst
@@ -4,11 +4,20 @@
 Change Log
 ****************
 
+- Improvement: DYNAMITE now works with newer versions of Astropy. The new requirement is astropy>=5.0.4
 - New feature: Integrate tube and box orbits in parallel by setting the multiprocessing option orblibs_in_parallel
+- New feature: Added support for new dark halo component type NFW_m200_c (fixed m200_c relation)
+- New feature: The Configuration object parameter reset_existing_output will delete previously existing data and create a new output directory tree
+- Improvement: The presence of datfil/orblib.dat.bz2 and datfil/orblibbox.dat.bz2 is now a more reliable indicator for existing orblibs. In the past, a crash may have resulted in corrupt bz2 files.
+- New feature: add new data-preparation method `BayesLOSVD.center_v_systemic`
+- New feature: Each model writes a file model_done_staging.ecsv upon completion. After a crash, DYNAMITE will update the all_models table with the completed models' data and delete any "all_done==False" models
+- New feature: New weight_solver_settings option reattempt_failures for reattempting failed weight solving when an orbit library already exists
+- Improvement: For better tracking, each model folder holds a copy of the config file now (instead of saving the config file in the output folder)
 - New feature: The new model iterator SplitModelIterator calculates orbit libraries and weights consecutively with independently adjustable number of threads
 - Improvement: Cleaned up the legacy_fortran folder and the makefile in it, unused orbgen.f90 and partgen.f90 moved to subfolder
 - Bugfix: Fixed a bug that on rare occasion caused an error when updating the timestamp entry when continuing an aborted run
 - Implement the correction to orbit mirroring introduced in `Quenneville et al 2021 <https://arxiv.org/abs/2111.06904>`_
+- Implement kinematic maps for BayesLOSVD data
 
 Version: 2.0
 ================